Refactor remote session lookup method
