+++ plugins/maya/publish/avalon_unlock_scene.py
@@ -4,6 +4,7 @@ import pyblish.api
 from maya import cmds
 from avalon import maya
 from reveries.maya import capsule
+from reveries import lib
 
 
 class AvalonUnlockScene(pyblish.api.ContextPlugin):
@@ -16,8 +17,7 @@ class AvalonUnlockScene(pyblish.api.ContextPlugin):
 
     def process(self, context):
 
-        if context.data.get("contractorAccepted"):
-            # Already in contractor's hand.
+        if lib.in_remote():
             return
 
         maya.unlock()
