+++ plugins/maya/publish/validate_scene_is_not_locked.py
@@ -1,6 +1,7 @@
 
 import pyblish.api
 from avalon import maya
+from reveries import lib
 
 
 class ValidateSceneNotLocked(pyblish.api.ContextPlugin):
@@ -13,7 +14,7 @@ class ValidateSceneNotLocked(pyblish.api.ContextPlugin):
 
     def process(self, context):
 
-        if context.data.get("contractorAccepted"):
+        if lib.in_remote():
             return
 
         if maya.is_locked():
